Fix SpotBugs issue and ensure test projects use current release.
SpotBugs requires a modern ASM release. Due to transitive dependencies
brought in via Spring's Dependency Management plugin ASM was being
fixed to 6.x during the SpotBugs task. This adds ASM to the dependency
management for the latest stable version (this will also matter when
updating to Groovy 2.5.7 to get Java 13 EA support).

Additionally, improves some issues with the tests. First, to catch any
subsequent SpotBugs issues the SpotBugs task is now run on the sample
projects during test jobs. Secondly, the ~/.gradle/caches copy of
all org.proticity.gradle plugins is wiped before running the tests.
Fetching a java-modern plugin from remote repos is banned, thus
requiring a copy from the local Maven repo, which has only the version
being currently built in it due to it being published there during the
job. Therefore if a change that bumps the version fails to also bump
the version number in the tests the tests will fail rather than
silently using an old version of the plugin. The modified caches are
not persisted since the cache policy for the test jobs is now pull
only.