Preparations for migrating to GitHub.
This addresses issues found in moving to GitHub that will help
any GitHub project, particularly for SonarQube integration.
For one, to support SonarCloud integration, it is easier to change
the SonarQube org. Previously it was locked to the Kordamp organization
info. Now it can be overridden by a sonar.organization property (not the
SonarQube plugin properties but normal Gradle properties), or the
SONAR_ORGANIZATION environment variable.

Also added support for CircleCI integration for determining the
branch name for SonarQube.