Remove JacocoFullReport task.
This plugin is no longer maintained and no longer works. It also tends
to break the regular Jacoco reports. It was being used primarily as a
premade JacocoMerge+Report task to feed a single report into SonarCloud
but it makes more sense to just provide all three individual reports and
let Sonar merge them. Where a coverage gate is desired for merged results
in a local build, this can be considered a future feature.